refactor(db): split AppDatabase DAOs into focused part files
Extract 12 DAO classes from app_database.dart (1058→190 lines) into
individual files under lib/data/db/daos/ as parts of the database
library. Recording helpers (recordingOverlapsEchoRegion,
_listEqualsRecordingRow) are co-located in recording_dao.dart.

- Each DAO lives in its own file with @DriftAccessor annotation
- DAO files are parts of the app_database library (no import changes
  needed for callers)
- AppDatabase migration helpers stay in app_database.dart
- All generated code remains in app_database.g.dart (structure
  unchanged by part-file extraction)
- flutter analyze: clean
- flutter test: 1331 passed, 0 failed

Closes #329
