+++ lib/data/db/app_database.dart
@@ -24,6 +24,18 @@ import 'tables/youtube_channel_subscriptions.dart';
 import 'tables/youtube_feed_entries.dart';
 
 part 'app_database.g.dart';
+part 'daos/ai_cache_dao.dart';
+part 'daos/audio_dao.dart';
+part 'daos/dictation_dao.dart';
+part 'daos/echo_session_dao.dart';
+part 'daos/recording_dao.dart';
+part 'daos/settings_dao.dart';
+part 'daos/sync_queue_dao.dart';
+part 'daos/transcript_dao.dart';
+part 'daos/transcript_fetch_state_dao.dart';
+part 'daos/video_dao.dart';
+part 'daos/youtube_channel_subscription_dao.dart';
+part 'daos/youtube_feed_entry_dao.dart';
 
 @DriftDatabase(
   tables: [
@@ -125,10 +137,6 @@ class AppDatabase extends _$AppDatabase {
       } else if (next == 11) {
         await _addColumnIfMissing(m, echoSessions, echoSessions.blurActive);
       } else if (next == 12) {
-        // Adds the L2 cache table for the AI result cache hierarchy
-        // (issue #311). Uses IF NOT EXISTS so a partial-migration crash
-        // on a previous launch cannot leave the app hanging on a blank
-        // window — mirroring the discipline of `_addColumnIfMissing`.
         await m.database.customStatement(
           'CREATE TABLE IF NOT EXISTS ai_cache ('
           'kind TEXT NOT NULL, '
@@ -191,868 +199,6 @@ class AppDatabase extends _$AppDatabase {
     ];
     for (final name in tables) {
       await m.database.customStatement('DROP TABLE IF EXISTS $name');
-    }
-  }
-}
-
-@DriftAccessor(tables: [Videos])
-class VideoDao extends DatabaseAccessor<AppDatabase> with _$VideoDaoMixin {
-  VideoDao(super.db);
-
-  Stream<List<VideoRow>> watchAll() => (select(
-    videos,
-  )..orderBy([(t) => OrderingTerm.desc(t.createdAt)])).watch();
-
-  Future<VideoRow?> getById(String id) =>
-      (select(videos)..where((t) => t.id.equals(id))).getSingleOrNull();
-
-  Future<VideoRow?> getYoutubeByVid(String youtubeVid) =>
-      (select(videos)..where(
-            (t) => t.provider.equals('youtube') & t.vid.equals(youtubeVid),
-          ))
-          .getSingleOrNull();
-
-  Future<List<VideoRow>> listAll() => select(videos).get();
-
-  Future<void> insertRow(VideoRow row) =>
-      into(videos).insert(row, mode: InsertMode.insertOrReplace);
-
-  /// Partial update so concurrent duration/thumbnail jobs do not clobber columns.
-  Future<void> updateLocalThumbnail(String id, String absoluteThumbPath) async {
-    await (update(videos)..where((t) => t.id.equals(id))).write(
-      VideosCompanion(
-        thumbnailUrl: Value(absoluteThumbPath),
-        updatedAt: Value(DateTime.now()),
-      ),
-    );
-  }
-
-  /// Partial update for YouTube oEmbed title/thumbnail refresh.
-  Future<void> updateYoutubeMetadata({
-    required String id,
-    required String title,
-    String? thumbnailUrl,
-  }) async {
-    await (update(videos)..where((t) => t.id.equals(id))).write(
-      VideosCompanion(
-        title: Value(title),
-        thumbnailUrl: thumbnailUrl == null
-            ? const Value.absent()
-            : Value(thumbnailUrl),
-        updatedAt: Value(DateTime.now()),
-      ),
-    );
-  }
-
-  Future<void> updateLanguage({
-    required String id,
-    required String language,
-  }) async {
-    await (update(videos)..where((t) => t.id.equals(id))).write(
-      VideosCompanion(
-        language: Value(language),
-        updatedAt: Value(DateTime.now()),
-      ),
-    );
-  }
-
-  Future<void> deleteId(String id) =>
-      (delete(videos)..where((t) => t.id.equals(id))).go();
-}
-
-@DriftAccessor(tables: [Audios])
-class AudioDao extends DatabaseAccessor<AppDatabase> with _$AudioDaoMixin {
-  AudioDao(super.db);
-
-  Stream<List<AudioRow>> watchAll() => (select(
-    audios,
-  )..orderBy([(t) => OrderingTerm.desc(t.createdAt)])).watch();
-
-  Future<AudioRow?> getById(String id) =>
-      (select(audios)..where((t) => t.id.equals(id))).getSingleOrNull();
-
-  /// Dedupe lookup by content hash (used by Craft from text import).
-  Future<AudioRow?> getByMd5(String md5) =>
-      (select(audios)..where((t) => t.md5.equals(md5))).getSingleOrNull();
-
-  Future<void> insertRow(AudioRow row) =>
-      into(audios).insert(row, mode: InsertMode.insertOrReplace);
-
-  Future<void> updateLanguage({
-    required String id,
-    required String language,
-  }) async {
-    await (update(audios)..where((t) => t.id.equals(id))).write(
-      AudiosCompanion(
-        language: Value(language),
-        updatedAt: Value(DateTime.now()),
-      ),
-    );
-  }
-
-  Future<void> deleteId(String id) =>
-      (delete(audios)..where((t) => t.id.equals(id))).go();
-}
-
-@DriftAccessor(tables: [Transcripts])
-class TranscriptDao extends DatabaseAccessor<AppDatabase>
-    with _$TranscriptDaoMixin {
-  TranscriptDao(super.db);
-
-  Stream<List<TranscriptRow>> watchForTarget(
-    String targetType,
-    String targetId,
-  ) =>
-      (select(transcripts)
-            ..where(
-              (t) =>
-                  t.targetType.equals(targetType) & t.targetId.equals(targetId),
-            )
-            ..orderBy([(t) => OrderingTerm.asc(t.language)]))
-          .watch();
-
-  Stream<List<TranscriptRow>> watchAllForTarget(
-    String targetType,
-    String targetId,
-  ) =>
-      (select(transcripts)
-            ..where(
-              (t) =>
-                  t.targetType.equals(targetType) & t.targetId.equals(targetId),
-            )
-            ..orderBy([
-              (t) => OrderingTerm.asc(t.source),
-              (t) => OrderingTerm.asc(t.language),
-              (t) => OrderingTerm.asc(t.createdAt),
-            ]))
-          .watch();
-
-  /// Existence-only stream (no `timeline_json` reads) for UI like transport buttons.
-  Stream<bool> watchExistsForTarget(String targetType, String targetId) {
-    return customSelect(
-      'SELECT EXISTS (SELECT 1 FROM transcripts WHERE target_type = ? AND target_id = ?) AS e',
-      variables: [
-        Variable.withString(targetType),
-        Variable.withString(targetId),
-      ],
-      readsFrom: {transcripts},
-    ).watch().map((rows) => rows.first.read<int>('e') != 0);
-  }
-
-  Future<TranscriptRow?> getById(String id) =>
-      (select(transcripts)..where((t) => t.id.equals(id))).getSingleOrNull();
-
-  Future<List<TranscriptRow>> listForTarget(
-    String targetType,
-    String targetId,
-  ) =>
-      (select(transcripts)..where(
-            (t) =>
-                t.targetType.equals(targetType) & t.targetId.equals(targetId),
-          ))
-          .get();
-
-  Future<void> upsert(TranscriptRow row) =>
-      into(transcripts).insert(row, mode: InsertMode.insertOrReplace);
-
-  Future<void> deleteId(String id) =>
-      (delete(transcripts)..where((t) => t.id.equals(id))).go();
-}
-
-@DriftAccessor(tables: [TranscriptFetchStates])
-class TranscriptFetchStateDao extends DatabaseAccessor<AppDatabase>
-    with _$TranscriptFetchStateDaoMixin {
-  TranscriptFetchStateDao(super.db);
-
-  Future<TranscriptFetchStateRow?> getForTarget(
-    String targetType,
-    String targetId,
-  ) =>
-      (select(transcriptFetchStates)..where(
-            (t) =>
-                t.targetType.equals(targetType) & t.targetId.equals(targetId),
-          ))
-          .getSingleOrNull();
-
-  Future<void> upsertFetched(
-    String targetType,
-    String targetId,
-    DateTime lastFetchedAt, {
-    String? lastStatus,
-    String? lastError,
-  }) => upsertOutcome(
-    targetType: targetType,
-    targetId: targetId,
-    lastFetchedAt: lastFetchedAt,
-    lastStatus: lastStatus ?? 'success',
-    lastError: lastError,
-  );
-
-  Future<void> upsertOutcome({
-    required String targetType,
-    required String targetId,
-    required DateTime lastFetchedAt,
-    required String lastStatus,
-    String? lastError,
-  }) => into(transcriptFetchStates).insert(
-    TranscriptFetchStateRow(
-      targetType: targetType,
-      targetId: targetId,
-      lastFetchedAt: lastFetchedAt,
-      lastStatus: lastStatus,
-      lastError: lastError,
-    ),
-    mode: InsertMode.insertOrReplace,
-  );
-
-  Future<void> clearForTarget(String targetType, String targetId) async {
-    await (delete(transcriptFetchStates)..where(
-          (t) => t.targetType.equals(targetType) & t.targetId.equals(targetId),
-        ))
-        .go();
-  }
-}
-
-@DriftAccessor(tables: [EchoSessions])
-class EchoSessionDao extends DatabaseAccessor<AppDatabase>
-    with _$EchoSessionDaoMixin {
-  EchoSessionDao(super.db);
-
-  // ignore: prefer_const_constructors — Uuid() is not const
-  static final Uuid _uuid = Uuid();
-
-  EchoSessionRow _newSession({
-    required String targetType,
-    required String targetId,
-    String language = 'und',
-    String? transcriptId,
-    String? secondaryTranscriptId,
-  }) {
-    final now = DateTime.now();
-    return EchoSessionRow(
-      id: _uuid.v4(),
-      targetType: targetType,
-      targetId: targetId,
-      language: language,
-      currentTimeMs: 0,
-      playbackRate: 1,
-      volume: 1,
-      echoStartMs: null,
-      echoEndMs: null,
-      transcriptId: transcriptId,
-      secondaryTranscriptId: secondaryTranscriptId,
-      recordingsCount: 0,
-      recordingsDurationMs: 0,
-      lastRecordingAt: null,
-      currentSegmentIndex: -1,
-      echoActive: false,
-      echoStartLine: -1,
-      echoEndLine: -1,
-      blurActive: false,
-      startedAt: now,
-      lastActiveAt: now,
-      completedAt: null,
-      syncStatus: null,
-      serverUpdatedAt: null,
-      createdAt: now,
-      updatedAt: now,
-    );
-  }
-
-  /// Ensures at least one echo session row exists for the target (creates if missing).
-  Future<EchoSessionRow> getOrCreateLatestForTarget(
-    String targetType,
-    String targetId,
-  ) async {
-    final existing = await getLatestForTarget(targetType, targetId);
-    if (existing != null) return existing;
-    final row = _newSession(targetType: targetType, targetId: targetId);
-    await into(echoSessions).insert(row);
-    return row;
-  }
-
-  Future<EchoSessionRow?> getLatestForTarget(
-    String targetType,
-    String targetId,
-  ) =>
-      (select(echoSessions)
-            ..where(
-              (t) =>
-                  t.targetType.equals(targetType) & t.targetId.equals(targetId),
-            )
-            ..orderBy([(t) => OrderingTerm.desc(t.lastActiveAt)])
-            ..limit(1))
-          .getSingleOrNull();
-
-  Stream<EchoSessionRow?> watchLatestForTarget(
-    String targetType,
-    String targetId,
-  ) =>
-      (select(echoSessions)
-            ..where(
-              (t) =>
-                  t.targetType.equals(targetType) & t.targetId.equals(targetId),
-            )
-            ..orderBy([(t) => OrderingTerm.desc(t.lastActiveAt)])
-            ..limit(1))
-          .watch()
-          .map((rows) => rows.isEmpty ? null : rows.first);
-
-  Future<void> upsert(EchoSessionRow row) =>
-      into(echoSessions).insert(row, mode: InsertMode.insertOrReplace);
-
-  Future<void> updatePrimaryTranscriptForTarget(
-    String targetType,
-    String targetId,
-    String? transcriptId,
-  ) async {
-    final latest = await getLatestForTarget(targetType, targetId);
-    final now = DateTime.now();
-    if (latest == null) {
-      await into(echoSessions).insert(
-        _newSession(
-          targetType: targetType,
-          targetId: targetId,
-          transcriptId: transcriptId,
-        ),
-      );
-    } else {
-      await (update(echoSessions)..where((t) => t.id.equals(latest.id))).write(
-        EchoSessionsCompanion(
-          transcriptId: Value(transcriptId),
-          updatedAt: Value(now),
-        ),
-      );
-    }
-  }
-
-  Future<void> updateSecondaryTranscriptForTarget(
-    String targetType,
-    String targetId,
-    String? secondaryTranscriptId,
-  ) async {
-    final latest = await getLatestForTarget(targetType, targetId);
-    final now = DateTime.now();
-    if (latest == null) {
-      await into(echoSessions).insert(
-        _newSession(
-          targetType: targetType,
-          targetId: targetId,
-          secondaryTranscriptId: secondaryTranscriptId,
-        ),
-      );
-    } else {
-      await (update(echoSessions)..where((t) => t.id.equals(latest.id))).write(
-        EchoSessionsCompanion(
-          secondaryTranscriptId: Value(secondaryTranscriptId),
-          updatedAt: Value(now),
-        ),
-      );
-    }
-  }
-
-  /// Local aggregates for profile stats (single round-trip).
-  Future<({int sessionCount, int recordingsDurationMs})> practiceTotals() {
-    return customSelect(
-          'SELECT COUNT(*) AS c, COALESCE(SUM(recordings_duration_ms), 0) AS d '
-          'FROM echo_sessions',
-          readsFrom: {echoSessions},
-        )
-        .map(
-          (row) => (
-            sessionCount: row.read<int>('c'),
-            recordingsDurationMs: row.read<int>('d'),
-          ),
-        )
-        .getSingle();
-  }
-}
-
-@DriftAccessor(tables: [Recordings])
-class RecordingDao extends DatabaseAccessor<AppDatabase>
-    with _$RecordingDaoMixin {
-  RecordingDao(super.db);
-
-  Stream<List<RecordingRow>> watchByTarget(
-    String targetType,
-    String targetId,
-  ) =>
-      (select(recordings)
-            ..where(
-              (t) =>
-                  t.targetType.equals(targetType) & t.targetId.equals(targetId),
-            )
-            ..orderBy([(t) => OrderingTerm.desc(t.createdAt)]))
-          .watch()
-          .distinctBy(_listEqualsRecordingRow);
-
-  /// Echo-region overlap — same rules as web `getRecordingsByEchoRegion`.
-  Stream<List<RecordingRow>> watchByEchoRegion({
-    required String targetType,
-    required String targetId,
-    required String language,
-    required int echoStartMs,
-    required int echoEndMs,
-  }) =>
-      (select(recordings)
-            ..where((t) {
-              final recordingEnd = t.referenceStart + t.referenceDuration;
-              return t.targetType.equals(targetType) &
-                  t.targetId.equals(targetId) &
-                  t.language.equals(language) &
-                  t.referenceStart.isSmallerThanValue(echoEndMs) &
-                  recordingEnd.isBiggerThanValue(echoStartMs);
-            })
-            ..orderBy([(t) => OrderingTerm.desc(t.createdAt)]))
-          .watch()
-          .distinctBy(_listEqualsRecordingRow);
-
-  Future<List<RecordingRow>> listByEchoRegion({
-    required String targetType,
-    required String targetId,
-    required String language,
-    required int echoStartMs,
-    required int echoEndMs,
-  }) async {
-    return (select(recordings)
-          ..where((t) {
-            final recordingEnd = t.referenceStart + t.referenceDuration;
-            return t.targetType.equals(targetType) &
-                t.targetId.equals(targetId) &
-                t.language.equals(language) &
-                t.referenceStart.isSmallerThanValue(echoEndMs) &
-                recordingEnd.isBiggerThanValue(echoStartMs);
-          })
-          ..orderBy([(t) => OrderingTerm.desc(t.createdAt)]))
-        .get();
-  }
-
-  Future<RecordingRow?> getById(String id) =>
-      (select(recordings)..where((t) => t.id.equals(id))).getSingleOrNull();
-
-  Future<void> insertRow(RecordingRow row) =>
-      into(recordings).insert(row, mode: InsertMode.insertOrReplace);
-
-  Future<void> updateAssessment({
-    required String id,
-    required int? pronunciationScore,
-    required String? assessmentJson,
-    required DateTime updatedAt,
-  }) => (update(recordings)..where((t) => t.id.equals(id))).write(
-    RecordingsCompanion(
-      pronunciationScore: Value(pronunciationScore),
-      assessmentJson: Value(assessmentJson),
-      updatedAt: Value(updatedAt),
-      syncStatus: const Value('local'),
-    ),
-  );
-
-  Future<void> deleteId(String id) =>
-      (delete(recordings)..where((t) => t.id.equals(id))).go();
-}
-
-/// Recording `[referenceStart, referenceStart + referenceDuration)` vs echo `[echoStartMs, echoEndMs)` (ms).
-bool recordingOverlapsEchoRegion(
-  RecordingRow r,
-  int echoStartMs,
-  int echoEndMs,
-) {
-  final recordingStart = r.referenceStart;
-  final recordingEnd = r.referenceStart + r.referenceDuration;
-  final overlapStart = recordingStart > echoStartMs
-      ? recordingStart
-      : echoStartMs;
-  final overlapEnd = recordingEnd < echoEndMs ? recordingEnd : echoEndMs;
-  return overlapStart < overlapEnd;
-}
-
-@DriftAccessor(tables: [Dictations])
-class DictationDao extends DatabaseAccessor<AppDatabase>
-    with _$DictationDaoMixin {
-  DictationDao(super.db);
-
-  Stream<List<DictationRow>> watchByTarget(
-    String targetType,
-    String targetId,
-  ) =>
-      (select(dictations)
-            ..where(
-              (t) =>
-                  t.targetType.equals(targetType) & t.targetId.equals(targetId),
-            )
-            ..orderBy([(t) => OrderingTerm.desc(t.createdAt)]))
-          .watch();
-
-  Future<void> insertRow(DictationRow row) =>
-      into(dictations).insert(row, mode: InsertMode.insertOrReplace);
-
-  Future<void> deleteId(String id) =>
-      (delete(dictations)..where((t) => t.id.equals(id))).go();
-}
-
-@DriftAccessor(tables: [SyncQueue])
-class SyncQueueDao extends DatabaseAccessor<AppDatabase>
-    with _$SyncQueueDaoMixin {
-  SyncQueueDao(super.db);
-
-  Future<int> enqueue({
-    required String entityType,
-    required String entityId,
-    required String action,
-    String? payloadJson,
-  }) => into(syncQueue).insert(
-    SyncQueueCompanion.insert(
-      entityType: entityType,
-      entityId: entityId,
-      action: action,
-      payloadJson: Value(payloadJson),
-      createdAt: DateTime.now(),
-    ),
-  );
-
-  Future<List<SyncQueueRow>> peekBatch({int limit = 50}) =>
-      (select(syncQueue)
-            ..orderBy([(t) => OrderingTerm.asc(t.createdAt)])
-            ..limit(limit))
-          .get();
-
-  Future<void> markAttempted(int id, {String? error}) async {
-    final existing = await (select(
-      syncQueue,
-    )..where((t) => t.id.equals(id))).getSingleOrNull();
-    if (existing == null) return;
-    await (update(syncQueue)..where((t) => t.id.equals(id))).write(
-      SyncQueueCompanion(
-        retryCount: Value(existing.retryCount + 1),
-        lastAttempt: Value(DateTime.now()),
-        error: Value(error),
-      ),
-    );
-  }
-
-  Future<void> deleteId(int id) =>
-      (delete(syncQueue)..where((t) => t.id.equals(id))).go();
-}
-
-@DriftAccessor(tables: [SettingsKv])
-class SettingsDao extends DatabaseAccessor<AppDatabase>
-    with _$SettingsDaoMixin {
-  SettingsDao(super.db);
-
-  static final _log = logNamed('SettingsDao');
-
-  void _assertKnownKey(String key) {
-    if (SettingsKeys.isKnown(key)) return;
-    final message = 'Unknown settings key: $key';
-    assert(() {
-      throw StateError(message);
-    }());
-    _log.warning(message);
-  }
-
-  Future<String?> getValue(String key) async {
-    _assertKnownKey(key);
-    final row = await (select(
-      settingsKv,
-    )..where((t) => t.key.equals(key))).getSingleOrNull();
-    return row?.value;
-  }
-
-  Future<void> setValue(String key, String value) {
-    _assertKnownKey(key);
-    return into(settingsKv).insert(
-      SettingRow(key: key, value: value),
-      mode: InsertMode.insertOrReplace,
-    );
-  }
-
-  /// Removes a key. No-op when the key is absent.
-  Future<void> deleteValue(String key) {
-    _assertKnownKey(key);
-    return (delete(settingsKv)..where((t) => t.key.equals(key))).go();
-  }
-}
-
-@DriftAccessor(tables: [YoutubeChannelSubscriptions])
-class YoutubeChannelSubscriptionDao extends DatabaseAccessor<AppDatabase>
-    with _$YoutubeChannelSubscriptionDaoMixin {
-  YoutubeChannelSubscriptionDao(super.db);
-
-  Stream<List<YoutubeChannelSubscriptionRow>> watchAll() => (select(
-    youtubeChannelSubscriptions,
-  )..orderBy([(t) => OrderingTerm.asc(t.displayName)])).watch();
-
-  Future<List<YoutubeChannelSubscriptionRow>> listAll() =>
-      select(youtubeChannelSubscriptions).get();
-
-  Future<YoutubeChannelSubscriptionRow?> getByChannelId(String channelId) =>
-      (select(
-        youtubeChannelSubscriptions,
-      )..where((t) => t.channelId.equals(channelId))).getSingleOrNull();
-
-  Future<void> upsert(YoutubeChannelSubscriptionRow row) => into(
-    youtubeChannelSubscriptions,
-  ).insert(row, mode: InsertMode.insertOrReplace);
-
-  Future<void> deleteChannelId(String channelId) => (delete(
-    youtubeChannelSubscriptions,
-  )..where((t) => t.channelId.equals(channelId))).go();
-
-  Future<void> touchLastFetched(String channelId, DateTime fetchedAt) async {
-    await (update(
-      youtubeChannelSubscriptions,
-    )..where((t) => t.channelId.equals(channelId))).write(
-      YoutubeChannelSubscriptionsCompanion(lastFetchedAt: Value(fetchedAt)),
-    );
-  }
-
-  Future<void> updateDisplayName(String channelId, String displayName) async {
-    await (update(
-      youtubeChannelSubscriptions,
-    )..where((t) => t.channelId.equals(channelId))).write(
-      YoutubeChannelSubscriptionsCompanion(displayName: Value(displayName)),
-    );
-  }
-
-  Future<void> updateThumbnail(String channelId, String? thumbnailUrl) async {
-    await (update(
-      youtubeChannelSubscriptions,
-    )..where((t) => t.channelId.equals(channelId))).write(
-      YoutubeChannelSubscriptionsCompanion(thumbnailUrl: Value(thumbnailUrl)),
-    );
-  }
-
-  Future<void> updateLanguage(String channelId, String language) async {
-    await (update(youtubeChannelSubscriptions)
-          ..where((t) => t.channelId.equals(channelId)))
-        .write(YoutubeChannelSubscriptionsCompanion(language: Value(language)));
-  }
-}
-
-@DriftAccessor(tables: [YoutubeFeedEntries])
-class YoutubeFeedEntryDao extends DatabaseAccessor<AppDatabase>
-    with _$YoutubeFeedEntryDaoMixin {
-  YoutubeFeedEntryDao(super.db);
-
-  Stream<List<YoutubeFeedEntryRow>> watchTimeline() => (select(
-    youtubeFeedEntries,
-  )..orderBy([(t) => OrderingTerm.desc(t.publishedAt)])).watch();
-
-  Stream<List<YoutubeFeedEntryRow>> watchForChannel(String channelId) =>
-      (select(youtubeFeedEntries)
-            ..where((t) => t.channelId.equals(channelId))
-            ..orderBy([(t) => OrderingTerm.desc(t.publishedAt)]))
-          .watch();
-
-  Future<void> upsertEntry(YoutubeFeedEntryRow row) =>
-      into(youtubeFeedEntries).insert(row, mode: InsertMode.insertOrReplace);
-
-  Future<YoutubeFeedEntryRow?> getEntry({
-    required String channelId,
-    required String videoId,
-  }) =>
-      (select(youtubeFeedEntries)..where(
-            (t) => t.channelId.equals(channelId) & t.videoId.equals(videoId),
-          ))
-          .getSingleOrNull();
-
-  Future<void> updateDurationSeconds({
-    required String channelId,
-    required String videoId,
-    required int durationSeconds,
-  }) async {
-    await (update(youtubeFeedEntries)..where(
-          (t) => t.channelId.equals(channelId) & t.videoId.equals(videoId),
-        ))
-        .write(
-          YoutubeFeedEntriesCompanion(durationSeconds: Value(durationSeconds)),
-        );
-  }
-
-  Future<void> deleteForChannel(String channelId) => (delete(
-    youtubeFeedEntries,
-  )..where((t) => t.channelId.equals(channelId))).go();
-}
-
-/// Element-wise comparison of two `RecordingRow` lists without allocating.
-///
-/// Used by `RecordingDao.watchByTarget` and `RecordingDao.watchByEchoRegion`
-/// to skip identical re-emissions — Drift re-queries on any change to the
-/// `recordings` table, and `shadow_reading_panel` plus
-/// `recordingsForTargetProvider` re-build on every emission.
-bool _listEqualsRecordingRow(
-  List<RecordingRow> previous,
-  List<RecordingRow> current,
-) {
-  if (identical(previous, current)) return true;
-  if (previous.length != current.length) return false;
-  for (var i = 0; i < previous.length; i++) {
-    final a = previous[i];
-    final b = current[i];
-    if (a.id != b.id ||
-        a.targetType != b.targetType ||
-        a.targetId != b.targetId ||
-        a.referenceStart != b.referenceStart ||
-        a.referenceDuration != b.referenceDuration ||
-        a.referenceText != b.referenceText ||
-        a.language != b.language ||
-        a.duration != b.duration ||
-        a.md5 != b.md5 ||
-        a.audioUrl != b.audioUrl ||
-        a.pronunciationScore != b.pronunciationScore ||
-        a.assessmentJson != b.assessmentJson ||
-        a.localPath != b.localPath ||
-        a.syncStatus != b.syncStatus ||
-        a.serverUpdatedAt != b.serverUpdatedAt ||
-        a.createdAt != b.createdAt ||
-        a.updatedAt != b.updatedAt) {
-      return false;
-    }
-  }
-  return true;
-}
-
-/// Drift accessor for the L2 tier of the AI result cache hierarchy.
-///
-/// All methods swallow Drift exceptions, log via `logNamed('ai_cache')`,
-/// and return null / no-op on failure. The cache layer never throws.
-@DriftAccessor(tables: [AiCache])
-class AiCacheDao extends DatabaseAccessor<AppDatabase> with _$AiCacheDaoMixin {
-  AiCacheDao(super.db);
-
-  static final _log = logNamed('ai_cache');
-
-  /// Single-row lookup. Returns null on miss or on Drift error.
-  Future<AiCacheRow?> read(String kind, String key) async {
-    try {
-      return await (select(aiCache)
-            ..where((t) => t.kind.equals(kind) & t.key.equals(key)))
-          .getSingleOrNull();
-    } on Object catch (e, st) {
-      _log.warning('ai_cache read failed kind=$kind key=$key', e, st);
-      return null;
-    }
-  }
-
-  /// Insert-or-replace. No-op on Drift error.
-  Future<void> upsert(
-    String kind,
-    String key,
-    String payloadJson,
-    DateTime updatedAt,
-  ) async {
-    try {
-      await into(aiCache).insert(
-        AiCacheRow(
-          kind: kind,
-          key: key,
-          payloadJson: payloadJson,
-          updatedAt: updatedAt.millisecondsSinceEpoch,
-        ),
-        mode: InsertMode.insertOrReplace,
-      );
-    } on Object catch (e, st) {
-      _log.warning('ai_cache upsert failed kind=$kind key=$key', e, st);
-    }
-  }
-
-  /// Single-row delete. No-op if missing or on Drift error.
-  Future<void> deleteRow(String kind, String key) async {
-    try {
-      await (delete(
-        aiCache,
-      )..where((t) => t.kind.equals(kind) & t.key.equals(key))).go();
-    } on Object catch (e, st) {
-      _log.warning('ai_cache deleteRow failed kind=$kind key=$key', e, st);
-    }
-  }
-
-  /// For [kind], keep the [keep] most-recent `updatedAt` rows; delete the
-  /// rest. Returns deleted count (best-effort; -1 on error).
-  Future<int> evictOldestExcept(String kind, int keep) async {
-    if (keep < 0) return 0;
-    try {
-      final total =
-          await (selectOnly(aiCache)
-                ..addColumns([aiCache.key.count()])
-                ..where(aiCache.kind.equals(kind)))
-              .map((row) => row.read<int>(aiCache.key.count()) ?? 0)
-              .getSingle();
-      if (total <= keep) return 0;
-      final toDelete = total - keep;
-      // Delete the oldest `toDelete` rows via a raw SQL DELETE in a
-      // subquery. SQLite supports DELETE ... WHERE key IN (SELECT ...).
-      final beforeKeys = await customSelect(
-        'SELECT key FROM ai_cache WHERE kind = ? '
-        'ORDER BY updated_at ASC LIMIT ?',
-        variables: [Variable.withString(kind), Variable.withInt(toDelete)],
-        readsFrom: {aiCache},
-      ).map((row) => row.read<String>('key')).get();
-      for (final k in beforeKeys) {
-        await (delete(
-          aiCache,
-        )..where((t) => t.kind.equals(kind) & t.key.equals(k))).go();
-      }
-      return beforeKeys.length;
-    } on Object catch (e, st) {
-      _log.warning('ai_cache evictOldestExcept failed kind=$kind', e, st);
-      return -1;
-    }
-  }
-
-  /// For [kind], delete every row whose `updatedAt < cutoff`. Returns
-  /// deleted count (best-effort; -1 on error).
-  Future<int> pruneOlderThan(String kind, DateTime cutoff) async {
-    try {
-      final before =
-          await (selectOnly(aiCache)
-                ..addColumns([aiCache.key.count()])
-                ..where(
-                  aiCache.kind.equals(kind) &
-                      aiCache.updatedAt.isSmallerThanValue(
-                        cutoff.millisecondsSinceEpoch,
-                      ),
-                ))
-              .map((row) => row.read<int>(aiCache.key.count()) ?? 0)
-              .getSingle();
-      await (delete(aiCache)..where(
-            (t) =>
-                t.kind.equals(kind) &
-                t.updatedAt.isSmallerThanValue(cutoff.millisecondsSinceEpoch),
-          ))
-          .go();
-      return before;
-    } on Object catch (e, st) {
-      _log.warning('ai_cache pruneOlderThan failed kind=$kind', e, st);
-      return -1;
-    }
-  }
-
-  /// Drops every row for [kind]. No-op on Drift error.
-  Future<void> deleteForKind(String kind) async {
-    try {
-      await (delete(aiCache)..where((t) => t.kind.equals(kind))).go();
-    } on Object catch (e, st) {
-      _log.warning('ai_cache deleteForKind failed kind=$kind', e, st);
-    }
-  }
-
-  /// Watch helper, used by tests and the diagnostic dump command.
-  Stream<List<AiCacheRow>> readAllForKind(String kind) {
-    return (select(aiCache)..where((t) => t.kind.equals(kind))).watch();
-  }
-
-  /// Synchronous row-count for [kind], used by `AiResultCache.stats()`.
-  Future<int> countForKind(String kind) async {
-    try {
-      return await (selectOnly(aiCache)
-            ..addColumns([aiCache.key.count()])
-            ..where(aiCache.kind.equals(kind)))
-          .map((row) => row.read<int>(aiCache.key.count()) ?? 0)
-          .getSingle();
-    } on Object catch (e, st) {
-      _log.warning('ai_cache countForKind failed kind=$kind', e, st);
-      return 0;
     }
   }
 }
