+++ lib/data/db/daos/video_dao.dart
@@ -1,0 +1,64 @@
+part of '../app_database.dart';
+
+@DriftAccessor(tables: [Videos])
+class VideoDao extends DatabaseAccessor<AppDatabase> with _$VideoDaoMixin {
+  VideoDao(super.db);
+
+  Stream<List<VideoRow>> watchAll() => (select(
+    videos,
+  )..orderBy([(t) => OrderingTerm.desc(t.createdAt)])).watch();
+
+  Future<VideoRow?> getById(String id) =>
+      (select(videos)..where((t) => t.id.equals(id))).getSingleOrNull();
+
+  Future<VideoRow?> getYoutubeByVid(String youtubeVid) =>
+      (select(videos)..where(
+            (t) => t.provider.equals('youtube') & t.vid.equals(youtubeVid),
+          ))
+          .getSingleOrNull();
+
+  Future<List<VideoRow>> listAll() => select(videos).get();
+
+  Future<void> insertRow(VideoRow row) =>
+      into(videos).insert(row, mode: InsertMode.insertOrReplace);
+
+  Future<void> updateLocalThumbnail(String id, String absoluteThumbPath) async {
+    await (update(videos)..where((t) => t.id.equals(id))).write(
+      VideosCompanion(
+        thumbnailUrl: Value(absoluteThumbPath),
+        updatedAt: Value(DateTime.now()),
+      ),
+    );
+  }
+
+  Future<void> updateYoutubeMetadata({
+    required String id,
+    required String title,
+    String? thumbnailUrl,
+  }) async {
+    await (update(videos)..where((t) => t.id.equals(id))).write(
+      VideosCompanion(
+        title: Value(title),
+        thumbnailUrl: thumbnailUrl == null
+            ? const Value.absent()
+            : Value(thumbnailUrl),
+        updatedAt: Value(DateTime.now()),
+      ),
+    );
+  }
+
+  Future<void> updateLanguage({
+    required String id,
+    required String language,
+  }) async {
+    await (update(videos)..where((t) => t.id.equals(id))).write(
+      VideosCompanion(
+        language: Value(language),
+        updatedAt: Value(DateTime.now()),
+      ),
+    );
+  }
+
+  Future<void> deleteId(String id) =>
+      (delete(videos)..where((t) => t.id.equals(id))).go();
+}
